% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_cpp.R
\name{bootstrap_cpp_integrated}
\alias{bootstrap_cpp_integrated}
\title{Bootstrap with cpp integration}
\usage{
bootstrap_cpp_integrated(
  X,
  B,
  dist = "NonParam",
  param1 = NULL,
  param2 = NULL,
  conf_level = 0.95
)
}
\arguments{
\item{X}{Original sample}

\item{B}{Number of replications}

\item{dist}{Specified distribution, allows for (non-)parametric estimation}

\item{param1}{Parameter of distribution in case of parametric estimation}

\item{param2}{Parameter of distribution in case of parametric estimation}

\item{conf_level}{Confidence levels for inference statistics}
}
\value{
An S3 object of class "bootstrap"
}
\description{
Bootstrap with cpp integration
}
