% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalProject-package.R
\docType{package}
\name{FinalProject-package}
\alias{FinalProject}
\alias{FinalProject-package}
\title{FinalProject: Bootstrapping}
\description{
This package implements several different versions of the bootstrap including; a regular R implementation, an integrated C++ version and a parallel computing version. These implementations fit different needs of bootstrapping depending on the amount of desired replications.
}
\author{
\strong{Maintainer}: Jarno Sprenkels \email{j.sprenkels@student.maastrichtuniversity.nl}

}
\keyword{internal}
