% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{plot.bootstrap}
\alias{plot.bootstrap}
\title{Plot method for "bootstrap" objects}
\usage{
\method{plot}{bootstrap}(object, ...)
}
\arguments{
\item{object}{bootstrap object}

\item{...}{additional arguments}
}
\description{
Plot method for "bootstrap" objects
}
