% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{print.summary.bootstrap}
\alias{print.summary.bootstrap}
\title{Print method for "summary.bootstrap" objects}
\usage{
\method{print}{summary.bootstrap}(x, digits = 4, ...)
}
\arguments{
\item{...}{additional arguments}

\item{object}{bootstrap object}
}
\description{
Print method for "summary.bootstrap" objects
}
