% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{summary.bootstrap}
\alias{summary.bootstrap}
\title{Summary method for "bootstrap" objects}
\usage{
\method{summary}{bootstrap}(object, ...)
}
\arguments{
\item{object}{Bootstrap object}

\item{...}{Additional arguments}
}
\description{
Summary method for "bootstrap" objects
}
