% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap}
\usage{
bootstrap(
  X,
  B,
  dist = "NonParam",
  param1 = NULL,
  param2 = NULL,
  conf_level = 0.95
)
}
\arguments{
\item{X}{Original sample}

\item{B}{Number of replications}

\item{dist}{Specified distribution, allows for (non)-parametric estimation}

\item{param1}{Parameter of distribution in case of parametric estimation}

\item{param2}{Parameter of distribution in case of parametric estimation}

\item{conf_level}{Confidence levels for inference statistics}
}
\value{
An S3 object of class "bootstrap"
}
\description{
This function allows to user to create an S3 object which contains values
based on the (non-)parametric bootstrapping estimated.
}
\examples{
set.seed(67)
data <- rnorm(500)

# Non-parametric bootstrap

result <- bootstrap(data, B = 1000, dist = "NonParam")
print(result)
summary(result)
plot(result)

# Normal parametric bootstrap with estimated parameters
result_norm <- bootstrap(data, B = 1000, dist = "Norm",
                                        param1 = mean(data),
                                        param2 = sd(data))

}
