% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_cpp.R
\name{bootstrap_cpp_integrated}
\alias{bootstrap_cpp_integrated}
\title{Bootstrap with C++ integration}
\usage{
bootstrap_cpp_integrated(
  X,
  B,
  dist = "NonParam",
  param1 = NULL,
  param2 = NULL,
  conf_level = 0.95
)
}
\arguments{
\item{X}{Original sample}

\item{B}{Number of replications}

\item{dist}{Specified distribution, allows for (non-)parametric estimation
must be "NonParam", "Norm" or "Exp", default is "NonParam"}

\item{param1}{Parameter 1 of distribution in case of parametric estimation}

\item{param2}{Parameter 2 of distribution in case of parametric estimation}

\item{conf_level}{Confidence levels for inference statistics, default is 0.95}
}
\value{
An S3 object of class "bootstrap"
}
\description{
This function allows to user to create an S3 object which contains values
based on the (non-)parametric bootstrapping estimated. Compared to the
regular bootstrap function, this uses C++ integration in Rcpp for better
performance
}
\examples{
\donttest{
set.seed(67)
data <- rnorm(500)

# Non-parametric bootstrap

result <- bootstrap_cpp_integrated(data, B = 1000, dist = "NonParam")
print(result)
summary(result)
plot(result)

# Normal parametric bootstrap with estimated parameters
result_norm <- bootstrap_cpp_integrated(data, B = 1000, dist = "Norm",
                                        param1 = mean(data),
                                        param2 = sd(data))
}
}
