% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_par_cpp.R
\name{bootstrap_par_cpp}
\alias{bootstrap_par_cpp}
\title{Parallel Bootstrap with C++ Integration}
\usage{
bootstrap_par_cpp(
  X,
  B,
  dist = "NonParam",
  param1 = NULL,
  param2 = NULL,
  conf_level = 0.95,
  cores = NULL
)
}
\arguments{
\item{X}{Original sample (numeric vector)}

\item{B}{Total number of bootstrap replications}

\item{dist}{Distribution type: "NonParam", "Norm", or "Exp"}

\item{param1}{Parameter 1 (mean for Norm, rate for Exp)}

\item{param2}{Parameter 2 (sd for Norm, ignored for Exp)}

\item{conf_level}{Confidence level for intervals (default 0.95)}

\item{cores}{Number of CPU cores to use for parallel computing. Setting this as
"NULL" will reduce this function to "bootstrap_cpp_integrated(...)."}
}
\value{
An S3 object of class "bootstrap" with same structure as
\code{bootstrap_cpp_integrated}
}
\description{
This function allows to user to create an S3 object which contains values
based on the (non-)parametric bootstrapping estimated. Compared to the
regular bootstrap function, this uses C++ integration and parallel computing
for better performance
}
\examples{
\donttest{
set.seed(67)
data <- rnorm(100)

# Parallel bootstrap with 2 cores

result_par_cpp <- bootstrap_par_cpp(data, B = 500, dist = "NonParam", cores = 2)
summary(result_par_cpp)
}
}
