% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalProject-package.R
\docType{package}
\name{FinalProject-package}
\alias{FinalProject}
\alias{FinalProject-package}
\title{FinalProject: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jarno Sprenkels \email{j.sprenkels@student.maastrichtuniversity.nl}

}
\keyword{internal}
